backend/requirements.txt
fastapi
uvicorn
psycopg2-binary
redis
aioredis
